+++ src/modules/device/index.ts
@@ -5,6 +5,6 @@
  * directory of this source tree.
  */
 
-export { DeviceCapabilityModule } from './DeviceCapabilityModule';
+export { DeviceModule } from './DeviceModule';
 export { IsEsimSupportedResponseSchema, IsEsimSupportedResultSchema } from './schemas';
 export type { IsEsimSupportedResponse, IsEsimSupportedResult } from './types';